+++ common/linked/text_opinions/wrapper.py
@@ -1,7 +1,4 @@
 from arekit.common.linked.data import LinkedDataWrapper
-from arekit.common.opinions.base import Opinion
-from arekit.common.text_opinions.end_type import EntityEndType
-from arekit.common.text_opinions.helper import TextOpinionHelper
 from arekit.common.text_opinions.text_opinion import TextOpinion
 
 
@@ -20,23 +17,12 @@ class LinkedTextOpinionsWrapper(LinkedDataWrapper):
     def RelatedNewsID(self):
         return self.First.NewsID
 
-    @staticmethod
-    def _aggregate_by_first(item, label):
-        assert(isinstance(item, TextOpinion))
-
-        # TODO: to opinion converter func.
-        # TODO: Move from here, as this is a fundamential transformation.
-        # TODO: Maybe in opinion helper.
-
-        source = TextOpinionHelper.extract_entity_value(item, EntityEndType.Source)
-        target = TextOpinionHelper.extract_entity_value(item, EntityEndType.Target)
-
-        return Opinion(source_value=source,
-                       target_value=target,
-                       sentiment=label)
-
     def get_linked_label(self):
         return self.First.Sentiment
+
+    def _get_data_label(self, item):
+        assert(isinstance(item, TextOpinion))
+        return item.Sentiment
 
     def get_prior_opinion_by_index(self, index):
         assert(isinstance(index, int))