+++ contrib/networks/sample.py
@@ -6,22 +6,19 @@ import numpy as np
 from arekit.common.entities.base import Entity
 from arekit.common.experiment.scales.base import BaseLabelScaler
 from arekit.common.frames.collection import FramesCollection
-from arekit.common.parsed_news.base import ParsedNews
+from arekit.common.model.sample import InputSampleBase
 from arekit.common.parsed_news.term_position import TermPositionTypes
 from arekit.common.synonyms import SynonymsCollection
-from arekit.common.text_opinions.text_opinion import TextOpinion
-from arekit.common.text_opinions.end_type import EntityEndType
+from arekit.common.text_opinions.enums import EntityEndType
 from arekit.common.text_opinions.helper import TextOpinionHelper
-
+from arekit.common.text_opinions.text_opinion import TextOpinion
 from arekit.contrib.networks.context.configurations.base.base import DefaultNetworkConfig
 from arekit.contrib.networks.features.dist import DistanceFeatures
 from arekit.contrib.networks.features.frames import FrameFeatures
 from arekit.contrib.networks.features.inds import IndicesFeature
 from arekit.contrib.networks.features.pointers import PointersFeature
 from arekit.contrib.networks.features.utils import pad_right_or_crop_inplace
-
 from arekit.networks.embedding import indices
-from arekit.common.model.sample import InputSampleBase
 
 
 class InputSample(InputSampleBase):